Removed test file, barrier exit function now called waitAtExit for clarity, makefile does not have test file info, fixed World constructor to take proper datatype unsigned long int

+++ src/world.cpp
@@ -2,7 +2,7 @@
 #include "world.h"
 #include "barrier.h"
 
-World::World(unsigned int numStranded) :
+World::World(long unsigned int numStranded) :
 	dock(new Barrier),
 	everyoneInBoat(0),
 	numAtIsland(numStranded),
